+++ engine/src/main/kotlin/com/undercut/util/Area.kt
@@ -1,6 +1,6 @@
 package com.undercut.util
 
-import com.undercut.game.Tile
+import world.gregs.voidps.type.Tile
 
 data class Area(
     val minX: Short,
@@ -9,10 +9,10 @@ data class Area(
     val maxY: Short
 ) {
     constructor(t1: Tile, t2: Tile) : this(
-        minX = minOf(t1.x, t2.x),
-        maxX = maxOf(t1.x, t2.x),
-        minY = minOf(t1.y, t2.y),
-        maxY = maxOf(t1.y, t2.y)
+        minX = minOf(t1.x, t2.x).toShort(),
+        maxX = maxOf(t1.x, t2.x).toShort(),
+        minY = minOf(t1.y, t2.y).toShort(),
+        maxY = maxOf(t1.y, t2.y).toShort()
     )
 
     constructor(cornerX1: Int, cornerY1: Int, cornerX2: Int, cornerY2: Int) : this(
@@ -29,7 +29,7 @@ data class Area(
     fun toTiles(): Array<Tile> = buildList {
         for (x in minX..maxX) {
             for (y in minY..maxY) {
-                add(Tile(x.toShort(), y.toShort(), 0))
+                add(Tile(x, y, 0))
             }
         }
     }.toTypedArray()
